Actions:
Modified publicationAdapter and card_view_publication for load image of a user in a card.
Modified CommentAdapter.java, CommentActivity.java, activity_comment.xml for load user image in a card.
Added navigationBar in "sign In" an "Comment"
Added landscape of "sign in"
Modified landscape of  "login".
Added icons.
Added dependencies on build.gradle.